+++ lib/monko-core/src/commonMain/kotlin/bson/MonkoBson.kt
@@ -1,6 +1,7 @@
 package dev.petuska.monko.core.bson
 
 import dev.petuska.monko.core.ext.Bson
+import dev.petuska.monko.core.ext.Document
 import dev.petuska.monko.core.util.NativeCloseable
 import dev.petuska.monko.core.util.Proxy
 
@@ -10,4 +11,8 @@ public interface MonkoBson : Proxy<Bson>, Bson, NativeCloseable {
 
 public expect fun bsonOf(json: String): MonkoBson
 
-public val EmptyBson: Bson = bsonOf("{}")
+public fun documentOf(json: String): Document = bsonOf(json).toDocument()
+
+public expect fun MonkoBson.toDocument(): Document
+
+public val EmptyBson: MonkoBson = bsonOf("{}")