+++ lib/monko-core/src/commonMain/kotlin/domain/result/InsertOneResult.kt
@@ -1,0 +1,6 @@
+package dev.petuska.monko.core.domain.result
+
+public data class InsertOneResult(
+  val acknowledged: Boolean,
+  val insertedId: String?
+)