+++ lib/monko-core/src/jsMain/kotlin/ext/type/result.kt
@@ -1,0 +1,8 @@
+package dev.petuska.monko.core.ext.type
+
+import kotlin.js.Json
+
+public external interface InsertOneResult : Json {
+  public var acknowledged: Boolean? get() = definedExternally; set(_) = definedExternally
+  public var insertedId: dynamic get() = definedExternally; set(_) = definedExternally
+}