+++ lib/monko-core/src/jvmMain/kotlin/MonkoClientJvm.kt
@@ -2,15 +2,12 @@ package dev.petuska.monko.core
 
 import com.mongodb.reactivestreams.client.MongoClients
 import dev.petuska.monko.core.ext.MongoClient
+import dev.petuska.monko.core.util.MPPValue
+import dev.petuska.monko.core.util.invoke
 import dev.petuska.monko.core.util.noop
 
-public actual suspend fun MonkoClient(
-  connectionString: String,
-  connectionStringJS: String?,
-  connectionStringJVM: String?,
-  connectionStringNative: String?,
-): MonkoClient {
-  return MongoClients.create(connectionStringJVM ?: connectionString).let(::MonkoClientJvm)
+public actual suspend fun MonkoClient(connectionString: MPPValue<String>): MonkoClient {
+  return MongoClients.create(connectionString()).let(::MonkoClientJvm)
 }
 
 internal class MonkoClientJvm(override val source: MongoClient) : MonkoClient {