+++ lib/monko-core/src/nativeMain/kotlin/util/MPPValue.kt
@@ -1,0 +1,3 @@
+package dev.petuska.monko.core.util
+
+public actual operator fun <T : Any> MPPValue<T>.invoke(): T = native ?: default