+++ sandbox/src/main/kotlin/main.kt
@@ -1,17 +1,8 @@
 package local.sandbox
 
-import dev.petuska.monko.core.CoreLib
-import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.delay
 
-fun main() {
-  val core = CoreLib()
-  println(core.sampleApi().withPlatform())
-  runBlocking {
-    suspendingMain()
-  }
-}
-
-suspend fun suspendingMain() {
-  val core = CoreLib()
-  println(core.sampleSuspendApi().withPlatformSuspend())
+suspend fun main() {
+  delay(1_000)
+  println("Hi")
 }