+++ scripts/setupUbuntu.sh
@@ -3,6 +3,7 @@
 sudo -v
 sudo apt install git cmake libssl-dev libsasl2-dev libncurses5-dev libncursesw5-dev libtinfo5 tar wget -y
 sudo ln -s /usr/lib/libncurses.so.5 /usr/lib/libtinfo.so.5
+sudo ln -s /usr/lib/libncurses.so.5 /lib/libtinfo.so.5
 
 "$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)/.build.sh" \
   "-DENABLE_AUTOMATIC_INIT_AND_CLEANUP=OFF" \